Icon: Use PNG images or CSS mask with SVG and background-color instead of WebFont
Use PNG images or CSS mask with SVG and background-color instead of WebFont

When the color is white or black, we can just include a white or black PNG image.

Otherwise, we include use div with a background-color and an SVG mask. This approach is less compatible (doesn't work in EdgeHTML 17), so it's not preferred.